Refactor inline TUI to component system
