+++ docs/guides/ratatui-integration.md
@@ -19,27 +19,29 @@ requirements:
 | Session bootstrap + renderer ownership | `run_single_agent_loop_unified` spawns a `RatatuiSession` and wires an `AnsiRenderer` sink.【F:src/agent/runloop/unified/turn.rs†L443-L477】 | The session exposes a `RatatuiHandle` for output commands and an event receiver for user input. |
 | Streaming response rendering | `AnsiRenderer::stream_markdown_response` forwards segments to the ratatui sink while maintaining the persistent transcript.【F:vtcode-core/src/utils/ansi.rs†L254-L302】 | Streaming replacements are issued through `RatatuiCommand::ReplaceLast` with an explicit `RatatuiMessageKind` so each block retains its styling. |
 | Prompt indicator, banner, tool summary | Rendered through existing helpers that now call `AnsiRenderer::with_ratatui` so the UI and transcript stay in sync.【F:src/agent/runloop/unified/display.rs†L1-L32】【F:src/agent/runloop/ui.rs†L13-L79】 |
-| Chat input loop | Managed by the modular TUI package: `RatatuiLoop` owns the input buffer and command queue while `events::handle_event` translates keyboard and mouse input into `RatatuiEvent` notifications for the agent.【F:vtcode-core/src/ui/tui/state/mod.rs†L1018-L1180】【F:vtcode-core/src/ui/tui/state/mod.rs†L1740-L1763】【F:vtcode-core/src/ui/tui/events/mod.rs†L14-L197】 |
-| Transcript + context trimming | Conversation tracking, scroll offsets, and PTY anchoring live inside the TUI state module, and the renderer consumes that metadata to build the transcript viewport without touching the persistent log.【F:vtcode-core/src/ui/tui/state/mod.rs†L479-L556】【F:vtcode-core/src/ui/tui/state/mod.rs†L972-L1180】【F:vtcode-core/src/ui/tui/render/mod.rs†L177-L245】【F:vtcode-core/src/ui/tui/render/mod.rs†L471-L546】 |
+| Chat input loop | Managed by the component-based TUI: `App` converts termion keyboard events into prompt edits, submissions, and scroll actions that surface as `RatatuiEvent` notifications for the agent runtime.【F:vtcode-core/src/ui/tui/app.rs†L94-L163】 |
+| Transcript + context trimming | The `Transcript` component owns scroll offsets, indicator rendering, and wrapped line generation so inline updates never disturb the persisted log.【F:vtcode-core/src/ui/tui/components/transcript.rs†L25-L198】 |
+| Prompt indicator and placeholder | The `Prompt` component renders the inline prefix, placeholder hint, and cursor offsets while applying theme colors from `RatatuiCommand::SetPrompt` and `SetPlaceholder`.【F:vtcode-core/src/ui/tui/components/prompt.rs†L102-L199】 |
 | Tool shell behavior | Unchanged: helpers render into the shared `AnsiRenderer` sink, which writes both to stdout and the ratatui session.【F:src/agent/runloop/unified/shell.rs†L1-L119】 |
 
 ## Key components introduced for `ratatui`
 
 1. **`vtcode-core/src/ui/tui.rs`**
-   - Owns session orchestration, including terminal setup, the async event loop, and the public `RatatuiSession`/`RatatuiHandle` API surface.【F:vtcode-core/src/ui/tui.rs†L1-L134】
-   - Routes drawing to the internal state machine while supervising autoresize handling and the periodic redraw ticker.【F:vtcode-core/src/ui/tui.rs†L39-L127】
+   - Exposes the session orchestration API, wiring configuration into `run_tui` and returning the `RatatuiSession`/`RatatuiHandle` pair used by the agent runtime.【F:vtcode-core/src/ui/tui.rs†L1-L49】
 
-2. **`vtcode-core/src/ui/tui/state`**
-   - Defines the shared data structures (`RatatuiCommand`, `RatatuiEvent`, scroll state, PTY tracking, status bar content) and encapsulates transcript mutation logic.【F:vtcode-core/src/ui/tui/state/mod.rs†L65-L156】【F:vtcode-core/src/ui/tui/state/mod.rs†L479-L1180】
-   - Provides helpers for draining command queues, applying themes, managing placeholders, and coordinating PTY output with transcript autoscroll.【F:vtcode-core/src/ui/tui/state/mod.rs†L1018-L1180】【F:vtcode-core/src/ui/tui/state/mod.rs†L1700-L1763】
+2. **`vtcode-core/src/ui/tui/tui.rs`**
+   - Builds the termion-backed terminal, spawns the blocking input listener, drains command queues, and schedules redraws within the inline viewport loop.【F:vtcode-core/src/ui/tui/tui.rs†L1-L162】
 
-3. **`vtcode-core/src/ui/tui/events`**
-   - Converts `crossterm` keyboard and mouse events into high-level actions, handles conversation navigation, and maintains slash-command selection state.【F:vtcode-core/src/ui/tui/events/mod.rs†L14-L197】
+3. **`vtcode-core/src/ui/tui/app.rs`**
+   - Maintains prompt and transcript state, applies `RatatuiCommand`s, and translates user actions into outbound `RatatuiEvent`s for the agent loop.【F:vtcode-core/src/ui/tui/app.rs†L14-L163】
 
-4. **`vtcode-core/src/ui/tui/render`**
-   - Implements layout calculation, transcript rendering, PTY window placement, input widgets, and the status bar using Ratatui widgets and styling helpers.【F:vtcode-core/src/ui/tui/render/mod.rs†L1-L245】【F:vtcode-core/src/ui/tui/render/mod.rs†L471-L546】
+4. **`vtcode-core/src/ui/tui/components/transcript.rs` & `components/prompt.rs`**
+   - Provide focused components for rendering the chat history and prompt line, encapsulating cursor math, wrapping, and indicator styling in a reusable layer.【F:vtcode-core/src/ui/tui/components/transcript.rs†L25-L198】【F:vtcode-core/src/ui/tui/components/prompt.rs†L102-L199】
 
-5. **`AnsiRenderer::with_ratatui`**
+5. **`vtcode-core/src/ui/tui/types.rs`**
+   - Declares the shared command/event types, text styling primitives, and session handle API used by both the renderer and higher-level run loop.【F:vtcode-core/src/ui/tui/types.rs†L1-L195】
+
+6. **`AnsiRenderer::with_ratatui`**
    - Wraps a `RatatuiHandle` so all structured output flows through ratatui while continuing to append
      to the persistent transcript log.【F:vtcode-core/src/utils/ansi.rs†L61-L155】
    - Converts `anstyle::Style` data into `ratatui` colors and modifiers to preserve syntax
