+++ vtcode-core/Cargo.toml
@@ -89,10 +89,6 @@ syntect = "5.2"
 ratatui = { version = "0.29", default-features = false, features = [
     "termion",
 ] }
-tuirealm = { version = "3.1.0", default-features = false, features = [
-    "derive",
-    "termion",
-] }
 unicode-width = "0.1"
 termion = "4.0"
 ignore = "0.4"
