+++ vtcode-core/src/ui/tui/components/mod.rs
@@ -1,0 +1,5 @@
+mod prompt;
+mod transcript;
+
+pub use prompt::Prompt;
+pub use transcript::Transcript;
