+++ vtcode-core/src/ui/tui/types.rs
@@ -1,0 +1,195 @@
+use ratatui::style::Color;
+use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender};
+
+#[derive(Clone, Default, PartialEq)]
+pub struct RatatuiTextStyle {
+    pub color: Option<Color>,
+    pub bold: bool,
+    pub italic: bool,
+}
+
+impl RatatuiTextStyle {
+    #[must_use]
+    pub fn merge_color(mut self, fallback: Option<Color>) -> Self {
+        if self.color.is_none() {
+            self.color = fallback;
+        }
+        self
+    }
+
+    #[must_use]
+    pub fn to_style(&self, fallback: Option<Color>) -> ratatui::style::Style {
+        let mut style = ratatui::style::Style::default();
+        if let Some(color) = self.color.or(fallback) {
+            style = style.fg(color);
+        }
+        if self.bold {
+            style = style.add_modifier(ratatui::style::Modifier::BOLD);
+        }
+        if self.italic {
+            style = style.add_modifier(ratatui::style::Modifier::ITALIC);
+        }
+        style
+    }
+}
+
+#[derive(Clone, Default)]
+pub struct RatatuiSegment {
+    pub text: String,
+    pub style: RatatuiTextStyle,
+}
+
+#[derive(Clone, Default)]
+pub struct RatatuiTheme {
+    pub background: Option<Color>,
+    pub foreground: Option<Color>,
+    pub primary: Option<Color>,
+    pub secondary: Option<Color>,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum RatatuiMessageKind {
+    Agent,
+    Error,
+    Info,
+    Policy,
+    Pty,
+    Tool,
+    User,
+}
+
+pub enum RatatuiCommand {
+    AppendLine {
+        kind: RatatuiMessageKind,
+        segments: Vec<RatatuiSegment>,
+    },
+    Inline {
+        kind: RatatuiMessageKind,
+        segment: RatatuiSegment,
+    },
+    ReplaceLast {
+        count: usize,
+        kind: RatatuiMessageKind,
+        lines: Vec<Vec<RatatuiSegment>>,
+    },
+    SetPrompt {
+        prefix: String,
+        style: RatatuiTextStyle,
+    },
+    SetPlaceholder {
+        hint: Option<String>,
+        style: Option<RatatuiTextStyle>,
+    },
+    SetMessageLabels {
+        agent: Option<String>,
+        user: Option<String>,
+    },
+    SetTheme {
+        theme: RatatuiTheme,
+    },
+    SetCursorVisible(bool),
+    SetInputEnabled(bool),
+    ClearInput,
+    ForceRedraw,
+    Shutdown,
+}
+
+#[derive(Debug, Clone)]
+pub enum RatatuiEvent {
+    Submit(String),
+    Cancel,
+    Exit,
+    Interrupt,
+    ScrollLineUp,
+    ScrollLineDown,
+    ScrollPageUp,
+    ScrollPageDown,
+}
+
+#[derive(Clone)]
+pub struct RatatuiHandle {
+    pub(crate) sender: UnboundedSender<RatatuiCommand>,
+}
+
+impl RatatuiHandle {
+    pub fn append_line(&self, kind: RatatuiMessageKind, segments: Vec<RatatuiSegment>) {
+        let segments = if segments.is_empty() {
+            vec![RatatuiSegment::default()]
+        } else {
+            segments
+        };
+        let _ = self
+            .sender
+            .send(RatatuiCommand::AppendLine { kind, segments });
+    }
+
+    pub fn inline(&self, kind: RatatuiMessageKind, segment: RatatuiSegment) {
+        let _ = self.sender.send(RatatuiCommand::Inline { kind, segment });
+    }
+
+    pub fn replace_last(
+        &self,
+        count: usize,
+        kind: RatatuiMessageKind,
+        lines: Vec<Vec<RatatuiSegment>>,
+    ) {
+        let _ = self
+            .sender
+            .send(RatatuiCommand::ReplaceLast { count, kind, lines });
+    }
+
+    pub fn set_prompt(&self, prefix: String, style: RatatuiTextStyle) {
+        let _ = self
+            .sender
+            .send(RatatuiCommand::SetPrompt { prefix, style });
+    }
+
+    pub fn set_placeholder(&self, hint: Option<String>) {
+        self.set_placeholder_with_style(hint, None);
+    }
+
+    pub fn set_placeholder_with_style(
+        &self,
+        hint: Option<String>,
+        style: Option<RatatuiTextStyle>,
+    ) {
+        let _ = self
+            .sender
+            .send(RatatuiCommand::SetPlaceholder { hint, style });
+    }
+
+    pub fn set_message_labels(&self, agent: Option<String>, user: Option<String>) {
+        let _ = self
+            .sender
+            .send(RatatuiCommand::SetMessageLabels { agent, user });
+    }
+
+    pub fn set_theme(&self, theme: RatatuiTheme) {
+        let _ = self.sender.send(RatatuiCommand::SetTheme { theme });
+    }
+
+    pub fn set_cursor_visible(&self, visible: bool) {
+        let _ = self.sender.send(RatatuiCommand::SetCursorVisible(visible));
+    }
+
+    pub fn set_input_enabled(&self, enabled: bool) {
+        let _ = self.sender.send(RatatuiCommand::SetInputEnabled(enabled));
+    }
+
+    pub fn clear_input(&self) {
+        let _ = self.sender.send(RatatuiCommand::ClearInput);
+    }
+
+    pub fn force_redraw(&self) {
+        let _ = self.sender.send(RatatuiCommand::ForceRedraw);
+    }
+
+    pub fn shutdown(&self) {
+        let _ = self.sender.send(RatatuiCommand::Shutdown);
+    }
+}
+
+pub struct RatatuiSession {
+    pub handle: RatatuiHandle,
+    pub events: UnboundedReceiver<RatatuiEvent>,
+}
